fix for Jabber privacy lists editor
git-svn-id: http://svn.miranda-ng.org/main/trunk@4457 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/JabberG/src/jabber_privacy.cpp
@@ -1356,9 +1356,12 @@ void CJabberDlgPrivacyLists::CListResetOptions(HWND)
 
 void CJabberDlgPrivacyLists::CListFilter(HWND)
 {
-	for (HANDLE hContact = db_find_first(m_proto->m_szModuleName); hContact; hContact = db_find_next(hContact, m_proto->m_szModuleName))
-		if (HANDLE hItem = m_clcClist.FindContact(hContact))
-			m_clcClist.DeleteItem(hItem);
+	for (HANDLE hContact = db_find_first(); hContact; hContact = db_find_next(hContact)) {
+		char *proto = GetContactProto(hContact);
+		if ( !proto || lstrcmpA(proto, m_proto->m_szModuleName))
+			if (HANDLE hItem = m_clcClist.FindContact(hContact))
+				m_clcClist.DeleteItem(hItem);
+	}
 }
 
 bool CJabberDlgPrivacyLists::CListIsGroup(HANDLE hGroup)
